Adds ACT compatible artifact upload/download
Adds support for simulating artifact upload and download when running workflows with ACT.

This is achieved by creating custom actions that copy artifacts to and from a temporary directory. This allows ACT to simulate the behavior of GitHub Actions' artifact upload and download features, enabling more accurate local testing of workflows.

Signed-off-by: Banee Ishaque K <Baneeishaque@gmail.com>