Add workflow to build and push Docker image
Adds a GitHub Actions workflow that automates the process of building and pushing Docker images to GitHub Container Registry.

This workflow supports multi-platform builds and utilizes inputs for image name, tags, platforms, and directories for API and web source code. It also checks out separate repositories for the API and web components, modifies the web app's axios baseURL for production, and generates metadata for image tags and labels based on commit SHAs.

Signed-off-by: Banee Ishaque K <Baneeishaque@gmail.com>